Endpoints done, more or less. Still have DELETE to handle http error

+++ api/app.py
@@ -1,6 +1,8 @@
+import flask
 from sqlalchemy.exc import SQLAlchemyError
 from flask import Flask, request, jsonify
 from flask_sqlalchemy import SQLAlchemy
+
 
 import os
 
@@ -35,7 +37,7 @@ def get_device(id):
     try:
         device = nbdevices.query.get(id)
         if not device:
-            return f"Device with ID: {id} does not exist"
+            return f"Device with ID: {id} does not exist", 409
         del device.__dict__['_sa_instance_state']
         return jsonify(device.__dict__)
     except SQLAlchemyError as e:
@@ -44,14 +46,25 @@ def get_device(id):
 
 
 
-
-
-# working !
+# NOT working w error handling http !
 @app.route('/devices/<id>', methods=['DELETE'])
 def delete_device(id):
-    db.session.query(nbdevices).filter_by(id=id).delete()
-    db.session.commit()
-    return f"Device with ID: {id} Removed"
+    try:
+        test = db.session.query(nbdevices).filter_by(id=id)
+        if not test:
+            return f'Unfortunally can not delete {id} Device with ID: {id} does not exist', 409
+        else:
+            db.session.query(nbdevices).filter_by(id=id).delete()
+            db.session.commit()
+            return f' Deleted ID: {id}'
+
+    except SQLAlchemyError as ex:
+        error = str(ex.__dict__['orig'])
+        return error, 422
+
+
+
+
 
 
 # working!
@@ -64,7 +77,7 @@ def create_device():
         return body
     except SQLAlchemyError as e:
         error = str(e.__dict__['orig'])
-        return jsonify(error)
+        return error, 422
 
 
 @app.route('/devices', methods=['GET'])
